+++ backend/routes/goal_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, request, jsonify
+from flask import Blueprint, request, jsonify, session
 from models import db, Goal
 from schemas import GoalSchema
 
@@ -11,21 +11,30 @@ goals_schema = GoalSchema(many=True)
 
 @goal_bp.route('/', methods=['GET'])
 def get_goals():
-    goals = Goal.query.all()
+    if 'user_id' not in session:
+        return {"error": "Unauthorized"}, 401
+
+    goals = Goal.query.filter_by(user_id=session['user_id']).all()
     result = goals_schema.dump(goals)
     return result, 200
 
 @goal_bp.route('/<int:id>', methods=['GET'])
 def get_goal(id):
-    goal = Goal.query.get_or_404(id)
+    if 'user_id' not in session:
+        return {"error": "Unauthorized"}, 401
+
+    goal = Goal.query.filter_by(id=id, user_id=session['user_id']).first_or_404()
     result = goal_schema.dump(goal)
     return result, 200
 
 @goal_bp.route('/', methods=['POST'])
 def create_goal():
+    if 'user_id' not in session:
+        return {"error": "Unauthorized"}, 401
+
     data = request.get_json()
     validated_data = goal_schema.load(data)
-    new_goal = Goal(**validated_data)
+    new_goal = Goal(**validated_data, user_id=session['user_id'])
     db.session.add(new_goal)
     db.session.commit()
     result = goal_schema.dump(new_goal)
@@ -33,9 +42,12 @@ def create_goal():
 
 @goal_bp.route('/<int:id>', methods=['PATCH'])
 def update_goal(id):
+    if 'user_id' not in session:
+        return {"error": "Unauthorized"}, 401
+
     data = request.get_json()
     validated_data = goal_schema.load(data, partial=True)
-    goal = Goal.query.get_or_404(id)
+    goal = Goal.query.filter_by(id=id, user_id=session['user_id']).first_or_404()
     for field, value in validated_data.items():
         setattr(goal, field, value)
     db.session.commit()
@@ -43,7 +55,10 @@ def update_goal(id):
 
 @goal_bp.route('/<int:id>', methods=['DELETE'])
 def delete_goal(id):
-    goal = Goal.query.get_or_404(id)
+    if 'user_id' not in session:
+        return {"error": "Unauthorized"}, 401
+
+    goal = Goal.query.filter_by(id=id, user_id=session['user_id']).first_or_404()
     db.session.delete(goal)
     db.session.commit()
     return {}, 204
